src/hooks/useBatteryData.ts
import { useState, useEffect, useCallback } from 'react';
import { ChargingData } from '../types/battery';
import { fetchBatteryData } from '../services/batteryApi';

export const useBatteryData = () => {
  const [data, setData] = useState<ChargingData | null>(null);
  const [loading, setLoading] = useState(true);
  const [error, setError] = useState<Error | null>(null);

  const fetchData = useCallback(async () => {
    try {
      setLoading(true);
      const result = await fetchBatteryData();
      setData(result);
      setError(null);
    } catch (err) {
      setError(err instanceof Error ? err : new Error('Failed to fetch data'));
      setData(null);
    } finally {
      setLoading(false);
    }
  }, []);

  useEffect(() => {
    fetchData();
  }, [fetchData]);

  const refetch = useCallback(() => {
    fetchData();
  }, [fetchData]);

  return { data, loading, error, refetch };
};